requirements.txt
gdown==5.2.0
pytorch==2.0.1
numpy==1.25.2
tqdm==4.65.0
